feat(file search): refactoring
