+++ src/test/kotlin/ai/masaic/openresponses/api/service/LocalFileStorageServiceTest.kt
@@ -1,5 +1,6 @@
 package ai.masaic.openresponses.api.service
 
+import ai.masaic.openresponses.api.service.storage.LocalFileStorageService
 import ai.masaic.openresponses.api.utils.toFilePart
 import com.fasterxml.jackson.databind.ObjectMapper
 import kotlinx.coroutines.flow.toList
