+++ src/test/kotlin/ai/masaic/openresponses/tool/FileSearchToolTest.kt
@@ -1,6 +1,6 @@
 package ai.masaic.openresponses.tool
 
-import ai.masaic.openresponses.api.service.VectorStoreService
+import ai.masaic.openresponses.api.service.search.VectorStoreService
 import ai.masaic.openresponses.tool.mcp.MCPToolExecutor
 import ai.masaic.openresponses.tool.mcp.MCPToolRegistry
 import com.openai.models.responses.ResponseCreateParams
